Remove deprecated method from PlayAreaView and update PlayAreaPresenter
