+++ app/src/main/java/at/vunfer/openrealms/view/PlayAreaView.java
@@ -16,11 +16,8 @@ import java.util.logging.Logger;
 
 /** View class for the play area. */
 public class PlayAreaView extends FrameLayout {
-    private static final Logger LOGGER = Logger.getLogger(PlayAreaView.class.getName());
-    private static final String TAG = "PlayAreaView";
-
     private TextView textView;
-    private ArrayList<Card> cards = new ArrayList<>();
+
     /**
      * Constructor for PlayAreaView.
      *
@@ -49,44 +46,10 @@ public class PlayAreaView extends FrameLayout {
      */
     public PlayAreaView(@NonNull Context context, @Nullable AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
-        initView();
-    }
-
-    /** Initializes the view. */
-    private void initView() {
-        LayoutInflater.from(getContext()).inflate(R.layout.play_area_view, this, true);
+        LayoutInflater.from(context).inflate(R.layout.play_area_view, this);
         textView = findViewById(R.id.text_play_area_view);
     }
-
-    /**
-     * Sets the text of the view.
-     *
-     * @param text The text to be set.
-     */
     public void setText(String text) {
-        try {
-            textView.setText(text);
-        } catch (Exception e) {
-            LOGGER.log(Level.SEVERE, TAG + ": Error setting text: " + e.getMessage());
-        }
-    }
-
-    /**
-     * Adds a card to the list of cards.
-     *
-     * @param card The card to be added.
-     */
-    public void addCard(Card card) {
-        cards.add(card);
-    }
-
-    /**
-     * Updates the view with the given text.
-     *
-     * @param text The text to be set.
-     */
-    public void updateView(String text) {
-        setText(text);
-        invalidate(); // Forces a redraw of the view
+        textView.setText(text);
     }
 }
